+++ packages/browser/src/index.ts
@@ -6,8 +6,10 @@ import { loadExtensions } from "@moonlight-mod/core/extension/loader";
 import { MoonlightBranch, MoonlightNode } from "@moonlight-mod/types";
 import { getConfig, getConfigOption, getManifest, setConfigOption } from "@moonlight-mod/core/util/config";
 import { IndexedDB } from "@zenfs/dom";
-import { configure } from "@zenfs/core";
+import { configureSingle } from "@zenfs/core";
 import * as fs from "@zenfs/core/promises";
+import { NodeEventPayloads, NodeEventType } from "@moonlight-mod/types/core/event";
+import { createEventEmitter } from "@moonlight-mod/core/util/event";
 
 function getParts(path: string) {
   if (path.startsWith("/")) path = path.substring(1);
@@ -18,15 +20,9 @@ window._moonlightBrowserInit = async () => {
   delete window._moonlightBrowserInit;
 
   // Set up a virtual filesystem with IndexedDB
-  await configure({
-    mounts: {
-      "/": {
-        backend: IndexedDB,
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore tsc tweaking
-        storeName: "moonlight-fs"
-      }
-    }
+  await configureSingle({
+    backend: IndexedDB,
+    storeName: "moonlight-fs"
   });
 
   window.moonlightNodeSandboxed = {
@@ -116,6 +112,7 @@ window._moonlightBrowserInit = async () => {
     processedExtensions,
     nativesCache: {},
     isBrowser: true,
+    events: createEventEmitter<NodeEventType, NodeEventPayloads>(),
 
     version: MOONLIGHT_VERSION,
     branch: MOONLIGHT_BRANCH as MoonlightBranch,
@@ -147,6 +144,7 @@ window._moonlightBrowserInit = async () => {
     async writeConfig(newConfig) {
       await writeConfig(newConfig);
       config = newConfig;
+      this.events.dispatchEvent(NodeEventType.ConfigSaved, newConfig);
     }
   };