+++ packages/types/src/core/event.ts
@@ -1,3 +1,4 @@
+import { Config } from "../config";
 import { WebpackModuleFunc, WebpackRequireType } from "../discord";
 
 export interface MoonlightEventEmitter<EventId extends string = string, EventData = Record<EventId, any>> {
@@ -6,16 +7,24 @@ export interface MoonlightEventEmitter<EventId extends string = string, EventDat
   removeEventListener: <Id extends keyof EventData>(id: Id, cb: (data: EventData[Id]) => void) => void;
 }
 
-export enum EventType {
+export enum WebEventType {
   ChunkLoad = "chunkLoad",
   ExtensionLoad = "extensionLoad"
 }
 
-export type EventPayloads = {
-  [EventType.ChunkLoad]: {
+export type WebEventPayloads = {
+  [WebEventType.ChunkLoad]: {
     chunkId?: number[];
     modules: { [id: string]: WebpackModuleFunc };
     require?: (require: WebpackRequireType) => any;
   };
-  [EventType.ExtensionLoad]: string;
+  [WebEventType.ExtensionLoad]: string;
+};
+
+export enum NodeEventType {
+  ConfigSaved = "configSaved"
+}
+
+export type NodeEventPayloads = {
+  [NodeEventType.ConfigSaved]: Config;
 };